requirements.txt
--extra-index-url https://backend:dc222fa6-452f-4253-a2fd-a52e7ba2e7e9@nexus.sigame.com.br/repository/pypi/simple
etria-logger==0.1.10
heimdall-client==2.3.1
persephone-client==2.2.0

motor==3.0.0
flask==2.1.2
python-decouple==3.6
aioboto3==9.6.0
httpx==0.23.0

uvloop==0.17.0
uvicorn==0.20.0
asgiref==3.5.2
hypercorn==0.13.2